tl;dr Added flyout menu, installed webkit and babel, added multiple pages to the website
Added a menu button on the top left corner as well as a flyout menu
that comes out from the left side when the menu button is clicked.
I also added 2 additional pages to the website, which can be accessed
via the flyout menu.

Installed webkit and babel, tools used by node and react respectively.

Added two more pages to the website, they are basically a copy of the
first page, but with no article content. These can be accesed via the
new menu.